A simple, multi-purpose router

The Waysign class implements a high-performance, multipurpose router build on top of the path-tree library. A router associates filepath-like patterns with a piece of data for latter retrieval. Often that data is a function and the path to be matched against the pattern comes from a URL, but it can be anything, adapting to the need of the user.

Details

The path pattern supported by Waysign mirrors that of path-tree and while the full documentation can be found there, it will be briefly explained here.

A path pattern consist of zero, one, or more elements separated by / (always started by /). Each element can either be a literal or one of the following variable types:

:name matches a single path piece

:name? matches an optional path piece

:name+ or + matches one or more path pieces

:name* or * matches zero or more path pieces

A variable don't have to consume a full path element. E.g. you could have a path pattern like this: /date/:day-:month-:year which would match to paths such as /date/24-12-2025

Methods

Public methods

Waysign$new()

Waysign$format()

Waysign$add_path()

Waysign$find_object()

Waysign$remove_path()

Waysign$clone()

Method new()

Create a new Waysign router

Usage

Waysign$new()

Method format()

Pretty printing of the object

Usage

Waysign$format(...)

Arguments

...

ignored

Method add_path()

Add a new path to the router. See Details for allowed path syntax

Usage

Waysign$add_path(path, object)

Arguments

path

A string giving the path to add

object

An R object to be associated with the path.

Returns

The object, invisibly

Method find_object()

Search for a path in the router

Usage

Waysign$find_object(path)

Arguments

path

The path to search for

Returns

If no matching path is found then NULL, otherwise a list with the elements path giving the path pattern that was matched, object giving the object associated with the path, keys giving the name of the path parameters in the path element, and values giving the values of the keys based on the match.

Method remove_path()

Remove a path from the router. Due to the underlying implementation this causes a complete rebuild of the router

Usage

Waysign$remove_path(path)

Arguments

path

A string giving the path to remove

Returns

The object, invisibly

Method clone()

The objects of this class are cloneable with this method.

Usage

Waysign$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Adapted from path-tree docs
router <- Waysign$new()

router$add_path("/", 1)
router$add_path("/login", 2)
router$add_path("/signup", 3)
router$add_path("/settings", 4)
router$add_path("/settings/:page", 5)
router$add_path("/:user", 6)
router$add_path("/:user/:repo", 7)
router$add_path("/public/:any*", 8)
router$add_path("/:org/:repo/releases/download/:tag/:filename.:ext", 9)
router$add_path("/:org/:repo/tags/:day-:month-:year", 10)
router$add_path("/:org/:repo/actions/:name\\::verb", 11)
router$add_path("/:org/:repo/:page", 12)
router$add_path("/:org/:repo/*", 13)
router$add_path("/api/+", 14)

router$find_object("/")
#> $path
#> [1] "/"
#> 
#> $object
#> [1] 1
#> 
#> $params
#> named character(0)
#> 
router$find_object("/login")
#> $path
#> [1] "/login"
#> 
#> $object
#> [1] 2
#> 
#> $params
#> named character(0)
#> 
router$find_object("/settings/admin")
#> $path
#> [1] "/settings/:page"
#> 
#> $object
#> [1] 5
#> 
#> $params
#>    page 
#> "admin" 
#> 
router$find_object("/viz-rs")
#> $path
#> [1] "/:user"
#> 
#> $object
#> [1] 6
#> 
#> $params
#>     user 
#> "viz-rs" 
#> 
router$find_object("/viz-rs/path-tree")
#> $path
#> [1] "/:user/:repo"
#> 
#> $object
#> [1] 7
#> 
#> $params
#>        user        repo 
#>    "viz-rs" "path-tree" 
#> 
router$find_object("/rust-lang/rust-analyzer/tags/2022-09-12")
#> $path
#> [1] "/:org/:repo/tags/:day-:month-:year"
#> 
#> $object
#> [1] 10
#> 
#> $params
#>             org            repo             day           month            year 
#>     "rust-lang" "rust-analyzer"          "2022"            "09"            "12" 
#> 
router$find_object("/rust-lang/rust-analyzer/actions/ci:bench")
#> $path
#> [1] "/:org/:repo/actions/:name\\::verb"
#> 
#> $object
#> [1] 11
#> 
#> $params
#>             org            repo            name            verb 
#>     "rust-lang" "rust-analyzer"            "ci"         "bench" 
#> 
router$find_object("/rust-lang/rust-analyzer/stargazers")
#> $path
#> [1] "/:org/:repo/:page"
#> 
#> $object
#> [1] 12
#> 
#> $params
#>             org            repo            page 
#>     "rust-lang" "rust-analyzer"    "stargazers" 
#> 
router$find_object("/rust-lang/rust-analyzer/stargazers/404")
#> $path
#> [1] "/:org/:repo/*"
#> 
#> $object
#> [1] 13
#> 
#> $params
#>              org             repo               *1 
#>      "rust-lang"  "rust-analyzer" "stargazers/404" 
#> 
router$find_object("/public/js/main.js")
#> $path
#> [1] "/public/:any*"
#> 
#> $object
#> [1] 8
#> 
#> $params
#>          any 
#> "js/main.js" 
#> 
router$find_object("/api/v1")
#> $path
#> [1] "/api/+"
#> 
#> $object
#> [1] 14
#> 
#> $params
#>   +1 
#> "v1" 
#>